Update GitHub Actions checkout action to version 3 in build_and_release.yml
The code changes update the GitHub Actions checkout action from version 2 to version 3. This ensures that the latest features and improvements are used during the code checkout process.

Additionally, the code changes simplify the compilation of tex files to pdf by removing unnecessary conditional checks. The xelatex command is now executed twice for each file, once for auxiliary files and once for generating the table of contents (ToC) and final PDF.

Overall, these changes improve the efficiency and reliability of the build and release workflow.